promo menu, staff and staff roster minor change('main')
