--- src/Staff.java
@@ -1,3 +1,0 @@
-public class Staff{
-    
-}
